third_party/WebKit/LayoutTests/inspector-protocol/timeline/timeline-layout.html

<html>
<head>
<style>
.my-class {
    min-width: 100px;
    background-color: red;
}
</style>
<script type="text/javascript" src="../../http/tests/inspector-protocol/inspector-protocol-test.js"></script>
<script>

function testFunction()
{
    var div = document.querySelector("#myDiv");
    div.classList.add("my-class");
    div.offsetWidth;
}

function test()
{
    InspectorTest.eventHandler["Tracing.dataCollected"] = dataCollected;
    InspectorTest.eventHandler["Tracing.tracingComplete"] = tracingComplete;
    InspectorTest.sendCommand("Tracing.start", { "categories" : "disabled-by-default-devtools.timeline", "type": "" }, onStart);

    function onStart(response)
    {
        InspectorTest.log("Recording started");
        InspectorTest.sendCommand("Runtime.evaluate", { "expression": "testFunction()" }, didEvaluate);
    }

    function didEvaluate(response)
    {
        InspectorTest.sendCommand("Tracing.end", { }, onStop);
    }

    var devtoolsEvents = [];
    function dataCollected(reply)
    {
        var allEvents = reply.params.value;
        devtoolsEvents = devtoolsEvents.concat(allEvents.filter(function(e)
        {
            return e.cat === "disabled-by-default-devtools.timeline";
        }));
    }

    function tracingComplete(event)
    {
        InspectorTest.log("Tracing complete");

        var schedRecalc = findEvent("ScheduleStyleRecalculation", "I");
        var recalcBegin = findEvent("RecalculateStyles", "B");
        var recalcEnd = findEvent("RecalculateStyles", "E");
        InspectorTest.assertEquals(schedRecalc.args.frame, recalcBegin.args.frame, "RecalculateStyles frame");
        InspectorTest.assert(recalcEnd.args.elementCount > 0, "RecalculateStyles elementCount");

        var invalidate = findEvent("InvalidateLayout", "I");
        var layoutBegin = findEvent("Layout", "B");
        var layoutEnd = findEvent("Layout", "E");

        InspectorTest.assertEquals(recalcBegin.args.frame, invalidate.args.frame, "InvalidateLayout frame");

        var beginData = layoutBegin.args.beginData;
        InspectorTest.assertEquals(invalidate.args.frame, beginData.frame, "Layout frame");
        InspectorTest.assert(beginData.dirtyObjects > 0, "dirtyObjects");
        InspectorTest.assert(beginData.totalObjects > 0, "totalObjects");

        var endData = layoutEnd.args.endData;
        InspectorTest.assert(endData.rootNode > 0, "rootNode id");
        InspectorTest.assert(!!endData.root , "root quad");

        InspectorTest.log("SUCCESS: found all expected events.");
        InspectorTest.completeTest();
    }

    function findEvent(name, ph)
    {
        for (var i = 0; i < devtoolsEvents.length; i++) {
            var e = devtoolsEvents[i];
            if (e.name === name && e.ph === ph)
                return e;
        }
        throw new Error("Couldn't find event " + name + " / " + ph + "\n\n in " + JSON.stringify(devtoolsEvents, null, 2));
    }

    function onStop(response)
    {
        InspectorTest.log("Recording stopped");
    }
}
</script>
</head>
<body onLoad="runTest();">
<div id="myDiv">DIV</div>
</body>
</html>
