third_party/WebKit/LayoutTests/inspector-protocol/timeline/timeline-layout-expected.txt

DIV
Recording started
Tracing complete
RecalculateStyles frames match: true
RecalculateStyles elementCount > 0: true
InvalidateLayout frames match: true
Layout frames match: true
dirtyObjects > 0: true
totalObjects > 0: true
has rootNode id: true
has root quad: true
SUCCESS: found all expected events.
